Client to make use of REST API
