+++ api/api.go
@@ -7,7 +7,6 @@ import (
 	"net/http"
 
 	"github.com/docker/libnetwork"
-	"github.com/docker/libnetwork/netutils"
 	"github.com/gorilla/mux"
 )
 
@@ -125,24 +124,9 @@ func makeHandler(ctrl libnetwork.NetworkController, fct processor) http.HandlerF
 	}
 }
 
-/***********
- Resources
-************/
-
-// networkResource is the body of the "get network" http response message
-type networkResource struct {
-	Name      string
-	ID        string
-	Type      string
-	Endpoints []*endpointResource
-}
-
-// endpointResource is the body of the "get endpoint" http response message
-type endpointResource struct {
-	Name    string
-	ID      string
-	Network string
-}
+/*****************
+ Resource Builders
+******************/
 
 func buildNetworkResource(nw libnetwork.Network) *networkResource {
 	r := &networkResource{}
@@ -170,51 +154,9 @@ func buildEndpointResource(ep libnetwork.Endpoint) *endpointResource {
 	return r
 }
 
-/***********
- Body types
-************/
-
-// networkCreate is the expected body of the "create network" http request message
-type networkCreate struct {
-	Name        string
-	NetworkType string
-	Options     map[string]interface{}
-}
-
-// endpointCreate represents the body of the "create endpoint" http request message
-type endpointCreate struct {
-	Name         string
-	NetworkID    string
-	ExposedPorts []netutils.TransportPort
-	PortMapping  []netutils.PortBinding
-}
-
-// endpointJoin represents the expected body of the "join endpoint" or "leave endpoint" http request messages
-type endpointJoin struct {
-	ContainerID       string
-	HostName          string
-	DomainName        string
-	HostsPath         string
-	ResolvConfPath    string
-	DNS               []string
-	ExtraHosts        []endpointExtraHost
-	ParentUpdates     []endpointParentUpdate
-	UseDefaultSandbox bool
-}
-
-// EndpointExtraHost represents the extra host object
-type endpointExtraHost struct {
-	Name    string
-	Address string
-}
-
-// EndpointParentUpdate is the object carrying the information about the
-// endpoint parent that needs to be updated
-type endpointParentUpdate struct {
-	EndpointID string
-	Name       string
-	Address    string
-}
+/**************
+ Options Parser
+***************/
 
 func (ej *endpointJoin) parseOptions() []libnetwork.EndpointOption {
 	var setFctList []libnetwork.EndpointOption
